@@ -4,7 +4,7 @@
     "type": "library",
     "license": "MIT",
     "require": {
-        "php": "^7.2",
+        "php": "^7.4",
         "symfony/validator": "^3.4.31|^4.2|^5.0",
         "symfony/intl": "^3.4|^4.2|^5.0",
         "khanamiryan/qrcode-detector-decoder": "^1.0.3",
@@ -13,13 +13,14 @@
     },
     "require-dev": {
         "phpunit/phpunit": "^8.0|^9.0",
+        "dg/bypass-finals": "^1.3",
         "dms/phpunit-arraysubset-asserts": "^0.1|^0.2",
-        "symfony/css-selector": "^4.2",
-        "phpstan/phpstan": "^0.12.23",
+        "fpdf/fpdf": "^1.82",
         "friendsofphp/php-cs-fixer": "^2.16",
+        "phpstan/phpstan": "^0.12.23",
+        "symfony/css-selector": "^4.2",
         "symfony/var-dumper": "^5.1",
-        "tecnickcom/tcpdf": "^6.3.2",
-        "fpdf/fpdf": "^1.82"
+        "tecnickcom/tcpdf": "^6.3.2"
     },
     "suggest": {
         "tecnickcom/tcpdf": "Needed to create pdfs with TcPdfOutput",
