@@ -4,11 +4,14 @@ namespace Sprain\SwissQrBill\Constraint;
 
 use Symfony\Component\Validator\Constraint;
 
-class ValidCreditorInformationPaymentReferenceCombination extends Constraint
+/**
+ * @internal
+ */
+final class ValidCreditorInformationPaymentReferenceCombination extends Constraint
 {
-    public $message = 'The payment reference type "{{ referenceType }}" does not match with the iban type of "{{ iban }}".';
+    public string $message = 'The payment reference type "{{ referenceType }}" does not match with the iban type of "{{ iban }}".';
 
-    public function getTargets()
+    public function getTargets(): string
     {
         return self::CLASS_CONSTRAINT;
     }
