@@ -4,7 +4,10 @@ namespace Sprain\SwissQrBill\Constraint;
 
 use Symfony\Component\Validator\Constraint;
 
-class ValidCreditorReference extends Constraint
+/**
+ * @internal
+ */
+final class ValidCreditorReference extends Constraint
 {
-    public $message = 'The string "{{ string }}" is not a valid Creditor Reference.';
+    public string $message = 'The string "{{ string }}" is not a valid Creditor Reference.';
 }
