@@ -11,7 +11,8 @@ use Sprain\SwissQrBill\DataGroup\Element\CreditorInformation;
 use Sprain\SwissQrBill\DataGroup\Element\Header;
 use Sprain\SwissQrBill\DataGroup\Element\PaymentAmountInformation;
 use Sprain\SwissQrBill\DataGroup\Element\PaymentReference;
-use Sprain\SwissQrBill\DataGroup\Element\StructuredAddress;
+use Sprain\SwissQrBill\DataGroup\EmptyElement\EmptyAdditionalInformation;
+use Sprain\SwissQrBill\DataGroup\EmptyElement\EmptyAddress;
 use Sprain\SwissQrBill\DataGroup\QrCodeableInterface;
 use Sprain\SwissQrBill\Exception\InvalidQrBillDataException;
 use Sprain\SwissQrBill\QrCode\QrCode;
@@ -21,45 +22,28 @@ use Sprain\SwissQrBill\Validator\SelfValidatableTrait;
 use Symfony\Component\Validator\Constraints as Assert;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 
-class QrBill implements SelfValidatableInterface
+final class QrBill implements SelfValidatableInterface
 {
     use SelfValidatableTrait;
 
-    const ERROR_CORRECTION_LEVEL_HIGH = ErrorCorrectionLevel::HIGH;
-    const ERROR_CORRECTION_LEVEL_MEDIUM = ErrorCorrectionLevel::MEDIUM;
-    const ERROR_CORRECTION_LEVEL_LOW = ErrorCorrectionLevel::LOW;
-
     private const SWISS_CROSS_LOGO_FILE = __DIR__ . '/../assets/swiss-cross.optimized.png';
-
     private const PX_QR_CODE = 543;    // recommended 46x46 mm in px @ 300dpi – in pixel based outputs, the final image size may be slightly different, depending on the qr code contents
     private const PX_SWISS_CROSS = 83; // recommended 7x7 mm in px @ 300dpi
 
-    /** @var Header */
-    private $header;
-
-    /** @var CreditorInformation */
-    private $creditorInformation;
-
-    /** @var AddressInterface*/
-    private $creditor;
-
-    /** @var PaymentAmountInformation */
-    private $paymentAmountInformation;
-
-    /** @var AddressInterface*/
-    private $ultimateDebtor;
-
-    /** @var PaymentReference */
-    private $paymentReference;
-
-    /** @var AdditionalInformation */
-    private $additionalInformation;
-
+    private Header $header;
+    private ?CreditorInformation $creditorInformation = null;
+    private ?AddressInterface $creditor = null;
+    private ?PaymentAmountInformation $paymentAmountInformation = null;
+    private ?AddressInterface $ultimateDebtor = null;
+    private ?PaymentReference $paymentReference = null;
+    private ?AdditionalInformation $additionalInformation = null;
     /** @var AlternativeScheme[] */
-    private $alternativeSchemes = [];
+    private array $alternativeSchemes = [];
 
-    /** @var string  */
-    private $errorCorrectionLevel = self::ERROR_CORRECTION_LEVEL_MEDIUM;
+    private function __construct(Header $header)
+    {
+        $this->header = $header;
+    }
 
     public static function create(): self
     {
@@ -69,13 +53,10 @@ class QrBill implements SelfValidatableInterface
             Header::CODING_LATIN
         );
 
-        $qrBill = new self();
-        $qrBill->header = $header;
-
-        return $qrBill;
+        return new self($header);
     }
 
-    public function getHeader(): ?Header
+    public function getHeader(): Header
     {
         return $this->header;
     }
@@ -179,15 +160,8 @@ class QrBill implements SelfValidatableInterface
     }
 
     /**
-     * @deprecated Will be removed in v3. The specs require the error correction level to be fixed at medium.
+     * @throws InvalidQrBillDataException
      */
-    public function setErrorCorrectionLevel(string $errorCorrectionLevel): self
-    {
-        $this->errorCorrectionLevel = $errorCorrectionLevel;
-
-        return $this;
-    }
-
     public function getQrCode(): QrCode
     {
         if (!$this->isValid()) {
@@ -204,7 +178,7 @@ class QrBill implements SelfValidatableInterface
         $qrCode->setLogoPath(self::SWISS_CROSS_LOGO_FILE);
         $qrCode->setRoundBlockSize(true, \Endroid\QrCode\QrCode::ROUND_BLOCK_SIZE_MODE_ENLARGE);
         $qrCode->setMargin(0);
-        $qrCode->setErrorCorrectionLevel(new ErrorCorrectionLevel($this->errorCorrectionLevel));
+        $qrCode->setErrorCorrectionLevel(new ErrorCorrectionLevel(ErrorCorrectionLevel::MEDIUM));
 
         return $qrCode;
     }
@@ -215,11 +189,11 @@ class QrBill implements SelfValidatableInterface
             $this->getHeader(),
             $this->getCreditorInformation(),
             $this->getCreditor(),
-            new StructuredAddress(), # Placeholder for ultimateCreditor, which is currently not yet allowed to be used by the implementation guidelines
+            new EmptyAddress(), # Placeholder for ultimateCreditor, which is currently not yet allowed to be used by the implementation guidelines
             $this->getPaymentAmountInformation(),
-            $this->getUltimateDebtor() ?: new StructuredAddress(),
+            $this->getUltimateDebtor() ?: new EmptyAddress(),
             $this->getPaymentReference(),
-            $this->getAdditionalInformation() ?: new AdditionalInformation(),
+            $this->getAdditionalInformation() ?: new EmptyAdditionalInformation(),
             $this->getAlternativeSchemes()
         ];
 
