@@ -10,14 +10,13 @@ use Sprain\SwissQrBill\Validator\SelfValidatableTrait;
 use Symfony\Component\Validator\Constraints as Assert;
 use Symfony\Component\Validator\Mapping\ClassMetadata;
 
-class RfCreditorReferenceGenerator implements SelfValidatableInterface
+final class RfCreditorReferenceGenerator implements SelfValidatableInterface
 {
     use SelfValidatableTrait;
 
-    /** @var string */
-    protected $reference;
+    private string $reference;
 
-    public static function generate(string $reference) : string
+    public static function generate(string $reference): string
     {
         $generator = new self($reference);
 
@@ -29,7 +28,7 @@ class RfCreditorReferenceGenerator implements SelfValidatableInterface
         $this->reference = StringModifier::stripWhitespace($reference);
     }
 
-    public function doGenerate() : string
+    public function doGenerate(): string
     {
         if (!$this->isValid()) {
             throw new InvalidCreditorReferenceException(
