@@ -2,6 +2,7 @@
 
 namespace Sprain\Tests\SwissQrBill\Constraints;
 
+use DG\BypassFinals;
 use Sprain\SwissQrBill\Constraint\ValidCreditorInformationPaymentReferenceCombination;
 use Sprain\SwissQrBill\Constraint\ValidCreditorInformationPaymentReferenceCombinationValidator;
 use Sprain\SwissQrBill\DataGroup\Element\CreditorInformation;
@@ -9,7 +10,7 @@ use Sprain\SwissQrBill\DataGroup\Element\PaymentReference;
 use Sprain\SwissQrBill\QrBill;
 use Symfony\Component\Validator\Test\ConstraintValidatorTestCase;
 
-class ValidCreditorInformationPaymentReferenceCombinationTest extends ConstraintValidatorTestCase
+final class ValidCreditorInformationPaymentReferenceCombinationTest extends ConstraintValidatorTestCase
 {
     protected function createValidator()
     {
@@ -33,21 +34,19 @@ class ValidCreditorInformationPaymentReferenceCombinationTest extends Constraint
     /**
      * @dataProvider emptyQrBillMocksProvider
      */
-    public function testEmptyQrBillValuesAreValid($qrBillMock)
+    public function testEmptyQrBillValuesAreValid(QrBill $qrBillMock)
     {
         $this->validator->validate($qrBillMock, new ValidCreditorInformationPaymentReferenceCombination());
 
         $this->assertNoViolation();
     }
 
-    public function emptyQrBillMocksProvider()
+    public function emptyQrBillMocksProvider(): array
     {
+        BypassFinals::enable();
+
         return [
             [$this->getQrBillMock()],
-            [$this->getQrBillMock(
-                $this->getCreditorInformationMock(),
-                $this->getPaymentReferenceMock()
-            )],
             [$this->getQrBillMock(
                 $this->getCreditorInformationMock(),
                 null
@@ -55,29 +54,21 @@ class ValidCreditorInformationPaymentReferenceCombinationTest extends Constraint
             [$this->getQrBillMock(
                 null,
                 $this->getPaymentReferenceMock()
-            )],
-            [$this->getQrBillMock(
-                $this->getCreditorInformationMock('any-iban'),
-                $this->getPaymentReferenceMock()
-            )],
-            [$this->getQrBillMock(
-                $this->getCreditorInformationMock(),
-                $this->getPaymentReferenceMock(PaymentReference::TYPE_QR)
-            )],
+            )]
         ];
     }
 
     /**
      * @dataProvider validCombinationsQrBillMocksProvider
      */
-    public function testValidCombinations($qrBillMock)
+    public function testValidCombinations(QrBill $qrBillMock)
     {
         $this->validator->validate($qrBillMock, new ValidCreditorInformationPaymentReferenceCombination());
 
         $this->assertNoViolation();
     }
 
-    public function validCombinationsQrBillMocksProvider()
+    public function validCombinationsQrBillMocksProvider(): array
     {
         return [
             [$this->getQrBillMock(
@@ -98,7 +89,7 @@ class ValidCreditorInformationPaymentReferenceCombinationTest extends Constraint
     /**
      * @dataProvider invalidCombinationsQrBillMocksProvider
      */
-    public function testInvalidCombinations($qrBillMock)
+    public function testInvalidCombinations(QrBill $qrBillMock)
     {
         $this->validator->validate($qrBillMock, new ValidCreditorInformationPaymentReferenceCombination([
             'message' => 'myMessage',
@@ -110,7 +101,7 @@ class ValidCreditorInformationPaymentReferenceCombinationTest extends Constraint
             ->assertRaised();
     }
 
-    public function invalidCombinationsQrBillMocksProvider()
+    public function invalidCombinationsQrBillMocksProvider(): array
     {
         return [
             [$this->getQrBillMock(
@@ -128,7 +119,7 @@ class ValidCreditorInformationPaymentReferenceCombinationTest extends Constraint
         ];
     }
 
-    public function getQrBillMock($creditorInformation = null, $paymentReference = null)
+    public function getQrBillMock(?CreditorInformation $creditorInformation = null, ?PaymentReference $paymentReference = null)
     {
         $qrBill = $this->createMock(QrBill::class);
 
@@ -141,7 +132,7 @@ class ValidCreditorInformationPaymentReferenceCombinationTest extends Constraint
         return $qrBill;
     }
 
-    public function getCreditorInformationMock($iban = null, $containsQrIban = false)
+    public function getCreditorInformationMock(string $iban = '', bool $containsQrIban = false)
     {
         $creditorInformation = $this->createMock(CreditorInformation::class);
 
@@ -154,7 +145,7 @@ class ValidCreditorInformationPaymentReferenceCombinationTest extends Constraint
         return $creditorInformation;
     }
 
-    public function getPaymentReferenceMock($paymentReferenceType = null)
+    public function getPaymentReferenceMock(string $paymentReferenceType = '')
     {
         $paymentReference = $this->createMock(PaymentReference::class);
 
