@@ -5,19 +5,19 @@ namespace Sprain\Tests\SwissQrBill\DataGroup\Element;
 use PHPUnit\Framework\TestCase;
 use Sprain\SwissQrBill\DataGroup\Element\AdditionalInformation;
 
-class AdditionalInformationTest extends TestCase
+final class AdditionalInformationTest extends TestCase
 {
     /**
      * @dataProvider messageProvider
      */
-    public function testMessage($numberOfValidations, $value)
+    public function testMessage(int $numberOfValidations, ?string $value): void
     {
         $additionalInformation = AdditionalInformation::create($value);
 
         $this->assertSame($numberOfValidations, $additionalInformation->getViolations()->count());
     }
 
-    public function messageProvider()
+    public function messageProvider(): array
     {
         return [
             [0, '012345678901234567890123456'],
@@ -30,14 +30,14 @@ class AdditionalInformationTest extends TestCase
     /**
      * @dataProvider billInformationProvider
      */
-    public function testBillInformation($numberOfValidations, $value)
+    public function testBillInformation(int $numberOfValidations, ?string $value)
     {
         $additionalInformation = AdditionalInformation::create(null, $value);
 
         $this->assertSame($numberOfValidations, $additionalInformation->getViolations()->count());
     }
 
-    public function billInformationProvider()
+    public function billInformationProvider(): array
     {
         return [
             [0, '012345678901234567890123456'],
@@ -47,7 +47,7 @@ class AdditionalInformationTest extends TestCase
         ];
     }
 
-    public function testFormattedString()
+    public function testFormattedString(): void
     {
         $additionalInformation = AdditionalInformation::create('message');
         $this->assertSame("message", $additionalInformation->getFormattedString());
@@ -56,7 +56,7 @@ class AdditionalInformationTest extends TestCase
         $this->assertSame("message\nbillInformation", $additionalInformation->getFormattedString());
     }
 
-    public function testQrCodeData()
+    public function testQrCodeData(): void
     {
         $additionalInformation = AdditionalInformation::create('message', 'billInformation');
 
