@@ -5,19 +5,19 @@ namespace Sprain\Tests\SwissQrBill\DataGroup\Element;
 use PHPUnit\Framework\TestCase;
 use Sprain\SwissQrBill\DataGroup\Element\AlternativeScheme;
 
-class AlternativeSchemeTest extends TestCase
+final class AlternativeSchemeTest extends TestCase
 {
     /**
      * @dataProvider parameterProvider
      */
-    public function testParameter($numberOfValidations, $value)
+    public function testParameter(int $numberOfValidations, string $value): void
     {
         $alternativeScheme = AlternativeScheme::create($value);
 
         $this->assertSame($numberOfValidations, $alternativeScheme->getViolations()->count());
     }
 
-    public function parameterProvider()
+    public function parameterProvider(): array
     {
         return [
             [0, '1'],
