@@ -6,19 +6,19 @@ use DMS\PHPUnitExtensions\ArraySubset\ArraySubsetAsserts;
 use PHPUnit\Framework\TestCase;
 use Sprain\SwissQrBill\PaymentPart\Translation\Translation;
 
-class TranslationTest extends TestCase
+final class TranslationTest extends TestCase
 {
     use ArraySubsetAsserts;
 
     /**
      * @dataProvider allTranslationsProvider
      */
-    public function testAllByLanguage($locale, $subset)
+    public function testAllByLanguage(string $locale, array $subset): void
     {
         $this->assertArraySubset($subset, Translation::getAllByLanguage($locale));
     }
 
-    public function allTranslationsProvider()
+    public function allTranslationsProvider(): array
     {
         return [
             ['de', ['paymentPart' => 'Zahlteil']],
@@ -31,12 +31,12 @@ class TranslationTest extends TestCase
     /**
      * @dataProvider singleTranslationProvider
      */
-    public function testGet($locale, $key, $translation)
+    public function testGet(string $locale, string $key, string $translation)
     {
         $this->assertSame($translation, Translation::get($key, $locale));
     }
 
-    public function singleTranslationProvider()
+    public function singleTranslationProvider(): array
     {
         return [
             ['de', 'paymentPart', 'Zahlteil'],
